project/requirements.txt
asgiref==3.4.1
asttokens==2.0.5
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.8
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==3.2.9
django-filter==21.1
django-redis==5.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-util==1.0.11
drf-yasg==1.20.0
executing==0.8.2
idna==3.3
importlib-resources==5.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.2.1
lorem==0.1.1
MarkupSafe==2.0.1
packaging==21.3
psycopg2-binary==2.9.2
pure-eval==0.2.1
PyJWT==2.3.0
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
redis==4.0.2
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
swagger-spec-validator==2.7.4
uritemplate==4.1.1
urllib3==1.26.7
varname==0.8.1
wrapt==1.13.3
zipp==3.6.0
